requirements.txt
paramiko
streamlit
pandas
plotly
requests
python-dotenv
streamlit-autorefresh
watchdog